backend/vector-service/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu